HOCS-4277: make queue listener generic
This change makes the queue listener generic and run off the interface
method `getMessageType`. At present this defaults to the single
UKVI_COMPLAINT message type as that's the only one. This will
eventually be driven from the message type that the user specifies.
