+++ src/main/java/uk/gov/digital/ho/hocs/queue/common/MessageHandler.java
@@ -1,0 +1,5 @@
+package uk.gov.digital.ho.hocs.queue.common;
+
+public interface MessageHandler {
+    void handleMessage(String message, String messageId) throws Exception;
+}
