--- src/main/java/uk/gov/digital/ho/hocs/queue/ukvi/UKVIComplaintQueueListener.java
@@ -1,26 +1,0 @@
-package uk.gov.digital.ho.hocs.queue.ukvi;
-
-import org.springframework.cloud.aws.messaging.listener.SqsMessageDeletionPolicy;
-import org.springframework.cloud.aws.messaging.listener.annotation.SqsListener;
-import org.springframework.messaging.handler.annotation.Header;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UKVIComplaintQueueListener {
-
-    private final UKVIComplaintService ukviComplaintService;
-    private final UKVIComplaintValidator ukviComplaintValidator;
-
-    public UKVIComplaintQueueListener(UKVIComplaintService ukviComplaintService,
-                                      UKVIComplaintValidator ukviComplaintValidator) {
-        this.ukviComplaintService = ukviComplaintService;
-        this.ukviComplaintValidator = ukviComplaintValidator;
-    }
-
-    @SqsListener(value = "${aws.sqs.ukvi-complaint.url}", deletionPolicy = SqsMessageDeletionPolicy.ON_SUCCESS)
-    public void onComplaintEvent(String message, @Header("MessageId") String messageId) throws Exception {
-        ukviComplaintValidator.validate(message, messageId);
-        ukviComplaintService.createComplaint(message, messageId);
-    }
-
-}
